requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.2
dynolayer==0.3.4
python-dotenv==1.0.1